# perf(mito): Use a heap to merge batches for the same key