I hereby agree to the terms of the [GreptimeDB CLA](https://gist.github.com/xtang/6378857777706e568c1949c7578592cc)

## What's changed and what's your intention?
This PR uses a heap to merge sorted batches in the `BatchMerger` instead of concatenating these batches and uses `sort_and_dedup()`. It also avoids concatenating batches as it introduces additional overhead.

These changes can speed up the merge reader when input batches don't have too many overlapping rows.

On Apple M1 Pro, merging 148797 sorted batches with 100,000,000+ rows took 2.2s before and 358ms now. Note that this only considers the cost of `BatchMerger`, not the whole `MergeReader`.

|  Before  | After |
|  ----  | ----  |
| 2.2s | 358ms |

It also fixes an issue that if a batch is empty after `filter_deleted()`, the reader will return `None` even if there are remaining batches.

## Checklist

- [ ]  I have written the necessary rustdoc comments.
- [ ]  I have added the necessary unit tests and integration tests.

## Refer to a related PR or issue link (optional)